refactor: Update DTO imports in controller tests for better organization
