+++ Backend/src/test/java/com/pg/mbti/RegistrationControllerTest.java
@@ -1,8 +1,8 @@
 package com.pg.mbti;
 
 import com.pg.mbti.controllers.RegistrationController;
-import com.pg.mbti.dto.RegistrationRequestDto;
-import com.pg.mbti.dto.RegistrationResponseDto;
+import com.pg.mbti.dto.auth.RegistrationRequestDto;
+import com.pg.mbti.dto.auth.RegistrationResponseDto;
 import com.pg.mbti.entity.User;
 import com.pg.mbti.enums.Gender;
 import com.pg.mbti.enums.MBTIType;
